Vital subsystem stable and returning expected values.

+++ trunk/PVMSBedsideMonitor/src/vital/BloodPressure.java
@@ -14,13 +14,13 @@ public class BloodPressure implements VitalAlgorithm {
 	VitalSensor sensor = new VitalSensor("bloodpressure");
 	
 	@Override
-	public int getValue(int rawMeasure) {
-		int realValue = MIN + ((MAX - MIN) * (rawMeasure/100) );
+	public int getValue(double rawMeasure) {
+		int realValue = (int) (MIN + ((MAX - MIN) * (rawMeasure/100) ));
 		return realValue;
 	}
 
 	public int getValue(){
-		int realValue = sensor.read();
+		double realValue = sensor.read();
 		return getValue(realValue);
 	}
