+++ trunk/PVMSBedsideMonitor/src/vital/VitalAlgorithm.java
@@ -8,6 +8,6 @@ public interface VitalAlgorithm {
 
 	//Get the raw (0-100) measurement and
 	//convert to the proper measure. 
-	int getValue(int rawMeasure);
+	int getValue(double rawMeasure);
 	int getValue();
 }
